refactor(accounting): re-land Wave-4 decompose P3-P6 (stacked merge dropped it onto parent branches)
#1202-1205 were marked MERGED but their content landed on the intermediate stacked
branches, NOT main (stacked PRs merged without retargeting bases to main). main still
had the 2,229-LOC AccountingService. This re-applies the full decompose onto current main
as a single base=main PR: facade (2,229→355 LOC) + PeakExportService/ReceivablesReportService/
TransactionalReportService/GeneralLedgerReportService. Behavior-preserving (already reviewed
in #1202-1205); gates green on current main (tsc 0, accounting 212/10, reports 65/6).
