+++ apps/api/src/modules/accounting/accounting.module.ts
@@ -2,6 +2,10 @@ import { Module } from '@nestjs/common';
 import { AccountingController } from './accounting.controller';
 import { AccountingClosingController } from './closing.controller';
 import { AccountingService } from './accounting.service';
+import { PeakExportService } from './peak-export.service';
+import { ReceivablesReportService } from './receivables-report.service';
+import { TransactionalReportService } from './transactional-report.service';
+import { GeneralLedgerReportService } from './general-ledger-report.service';
 import { BadDebtService } from './bad-debt.service';
 import { BadDebtProvisionCron } from './bad-debt-provision.cron';
 import { MonthlyCloseService } from './monthly-close.service';
@@ -18,6 +22,10 @@ import { PeakModule } from '../peak/peak.module';
   controllers: [AccountingController, AccountingClosingController, ConsolidatedController],
   providers: [
     AccountingService,
+    PeakExportService,
+    ReceivablesReportService,
+    TransactionalReportService,
+    GeneralLedgerReportService,
     BadDebtService,
     BadDebtProvisionCron,
     MonthlyCloseService,
