+++ apps/api/src/modules/accounting/accounting.pl-expense.spec.ts
@@ -3,6 +3,10 @@ import { PrismaService } from '../../prisma/prisma.service';
 import { JournalAutoService } from '../journal/journal-auto.service';
 import { CompanyResolverService } from '../journal/company-resolver.service';
 import { AccountingService } from './accounting.service';
+import { PeakExportService } from './peak-export.service';
+import { ReceivablesReportService } from './receivables-report.service';
+import { TransactionalReportService } from './transactional-report.service';
+import { GeneralLedgerReportService } from './general-ledger-report.service';
 
 type GroupRow = { accountCode: string; _sum: { debit: Prisma.Decimal | null; credit: Prisma.Decimal | null } };
 
@@ -16,11 +20,12 @@ function makeService(groupRows: GroupRow[]) {
   const companyResolver = {
     getFinanceCompanyId: jest.fn().mockResolvedValue('finance-co-1'),
   } as unknown as CompanyResolverService;
-  const svc = new AccountingService(prisma, {} as JournalAutoService, companyResolver);
+  // Wave-4 P5: aggregateFinanceExpenses now lives (private) on TransactionalReportService.
+  const svc = new TransactionalReportService(prisma, companyResolver);
   return { svc, journalLineGroupBy };
 }
 
-const aggregate = (svc: AccountingService, companyWide: boolean) =>
+const aggregate = (svc: TransactionalReportService, companyWide: boolean) =>
   (svc as unknown as {
     aggregateFinanceExpenses: (
       s: Date,
@@ -85,7 +90,18 @@ function makeFullService(groupRows: GroupRow[]) {
   const companyResolver = {
     getFinanceCompanyId: jest.fn().mockResolvedValue('finance-co-1'),
   } as unknown as CompanyResolverService;
-  return new AccountingService(prisma, {} as JournalAutoService, companyResolver);
+  return new AccountingService(
+    prisma,
+    {} as JournalAutoService,
+    companyResolver,
+    {} as PeakExportService,
+    {} as ReceivablesReportService,
+    // Wave-4 P5: real collaborator so getProfitLossReport delegates correctly.
+    new TransactionalReportService(prisma, companyResolver),
+    // Wave-4 P6: not exercised here (these tests only hit getProfitLossReport),
+    // so a stub satisfies the constructor arity.
+    {} as GeneralLedgerReportService,
+  );
 }
 
 describe('AccountingService.getProfitLossReport (expense wiring)', () => {
@@ -138,7 +154,18 @@ describe('AccountingService.getMonthlyPLSummary (expense wiring)', () => {
     const companyResolver = {
       getFinanceCompanyId: jest.fn().mockResolvedValue('finance-co-1'),
     } as unknown as CompanyResolverService;
-    return new AccountingService(prisma, {} as JournalAutoService, companyResolver);
+    return new AccountingService(
+      prisma,
+      {} as JournalAutoService,
+      companyResolver,
+      {} as PeakExportService,
+      {} as ReceivablesReportService,
+      // Wave-4 P5: real collaborator so getMonthlyPLSummary delegates correctly.
+      new TransactionalReportService(prisma, companyResolver),
+      // Wave-4 P6: not exercised here (these tests only hit getMonthlyPLSummary),
+      // so a stub satisfies the constructor arity.
+      {} as GeneralLedgerReportService,
+    );
   }
 
   it('includeFinanceExpenses=true: per-month FINANCE expenses subtracted from revenue', async () => {
